backend/api: Support HTTP DELETE on /api/v0/results
It's extremely convenient for the tests fixtures to delete all results
that were added during a unit test so that the next test to run has
a clean state. This change allows us to revert a workaround in
test_add_multiple_test_results_at_once() caused because of results added
in previous unit tests.