backend: Annotate change points with GitHub commit messages
For most devs, the git sha is basically meaningless and there's much
more signal in the first line of the commit message. So annotate the
change point attributes with this information and display it in the
dashboard change point table.

Additionally, since we know if a commit is hosted on github.com (thanks
to the attributes) we can link to the commit.