Allow users to sign up
Users can now sign up via the front page and will receive an email
asking them to verify their account.

The beginnings of an email template framework is included in this
commit. In the future we'll probably want to also use it for password
recovery emails.

This change adds a couple of new required environment variables:

 - POSTMARK_API_KEY -- API key for postmarkapp.com account
 - SERVER_NAME -- the protocol and hostname of the server, e.g. https://nyrk.io