frontend/dashboard: Don't display chart data points for lots of data
If we have many data points on a chart showing a little circle for the
data point becomes really chaotic and makes it difficult to see the line
anymore.

The original motivation for including points for non-change-point data
was that it can help to see the data points live if you have a lot of
variability in the data set. That doesn't really apply if you have many
points because things tend to smoothe out over time.