Provide default data for every new account
Provide a publicly accessible set of default data that anyone can access
without authorisation. These APIs are separate from the usual ones and
as such DELETE isn't supported on these endpoints (for obvious reasons).

We can also include this default data on the front page to give users a
taste of the service without needing to create an account.

Now that we're using the same code in multiple places (on the front page
and on the dashboard) let's also make the code more modular and create a
bunch of new react components in the frontend.