backend/core: Store series results in timestamp order
If we don't enforce an ordering when passing a test result series to the
change point detection algorithm then we'll get back nonsense results
because we're claiming the input is a time series when it's not.

Maintain the results in sorted order inside of
PerformanceTestResultSeries so that the callers don't need to worry
about this implementation detail.