db: Sort default data test results by timestamp
The core change detection algorithm requires that the time series is,
well, a time series. We already have this sort clause on the query for
the user test results but we need to duplicate it here too.

This isn't a correctness fix, the core already ensures that the time
series is maintained internally in sorted order. But it will unlock some
performance opportunities later, e.g. adding sorted results in batch.