core/github: Use SIEVE cache for GitHub API responses
We frequently need to fetch the same commit messages over and over (e.g.
the front page displays the tigerbeetle data) and we should cache
responses to avoid angering the GitHub rate limiter.

The expected maximum size of the cache (assuming the first line of a
typical commit message is no more than 80 characters) is 1.25MiB.