backend/api/config: Per-test config endpoint
This is groundwork to allow users to mark certain tests as public, i.e.
viewable by non-authenticated users. Each test config requires an
attributes dictionary that states the git repository and branch name the
config applies to, just like test results.

Unlike the test result schema, it is not possible to add arbitrary
fields to the config. The schema is enforced by the API layer.

A future addition would be to allow user's to configure per-test p-value
and magnitude settings too.