frontend/dashboard: Extract test name from url not state
When users visit a test page directly there is no location state
supplied since it comes from the <Link> react component. Instead of
relying on this state (which won't always be there) just extract the
test name from the url and handle the case where it's invalid and we hit
a 404.

Fixes #100