backend/api/pull_request: Change detection for GitHub pull requests
This API allows users to upload results from GitHub pull requests and
run change detection on them. A new API endpoint is required because A)
we need additional info about the test results (which PR it comes from)
and B) we need a way to separate regular test results from PR test
results.

Much like the organisation endpoint, the PR enpoint uses the GitHub
"org/repo" syntax for the repo name.

GitHub PR comments can also be sent to the PR by hitting the following
endpoint with the `notify` url parameter:

  /pulls/{repo}/{pr_number}/changes/{git_sha}?notify=1

Note that we don't currently support running change detection on
individual commits in a PR. Just the last commit.

As part of this API, we now need to deploy a public key infra file
(nyrkio.pem) which is used to sign the payload uploaded as part of the
PR comment.