certbot: Roll our own certbot image for ssl cert creation
We need a better way to deal with the chicken-and-egg problem of
certificate creation needing nginx to be running but nginx needing
certificates to run.

The canonical solution to this is to create dummy certificates just to
get nginx up and running so that real certs can be generated.

This also required making nginx's configuration use templates to inject
the domain name at build time.