backend: User result sharing via GitHub Organizations
If a user is authenticated via GitHub OAuth then we can track which
GitHub organizations they belong to where they've installed the Nyrkiö
GH app. This allows users in the same organization to share results.

This necessiates a new API endpoints at /api/v0/orgs which works
similarly to the other endpoints.