Get rid of public_results collection for public tests
Public tests (test results viewable by anyone regardless of
authentication) is currently managed via two collections: test_config
and public_results.

Keeping two separate collections in sync is a headache that is best
avoided, and we can figure out which tests are public just from their
test config. This simplifies things enormously.

At the same time, change the output of /public/results so clients do
not need to build urls anymore from the result attributes. Anything
coming out of the public results can be assumed to be a GitHub result.

All of this logic should really live in the public API code and not be
part of the DB layer. Move it.