auth: Allow users to generate new API tokens
Add an API for authenticated users to generate a new API token and
provide a new section on the user settings page.

The primary use case for this is users that logged in via GitHub. Such
users don't have a user/pass combination to provide to the existing
/api/v0/auth/jwt/login endpoint which is the usual way to get an access
token.

Fixes #126