backend/core: Temporarily stop GitHub commit fetch when hitting ratelimit
There's no sense in continuing to hammer the GitHub API if we know the
requests are going to return 403 (Forbidden) or 429 (Too many requests)
until some future time. Extract that time from the X-ratelimit-reset
response header and disable GitHub commit message fetching until
afterwards.

And added bonus is that we no longer flood the logs with "Failed to
fetch" commit messages which seems to affect request latency.