backend/core: Skip unit tests if GH commit fetch is paused
After 9290596 ("backend/core: Temporarily stop GitHub commit fetch when
hitting ratelimit") we now won't raise an exception if the GitHub
ratelimit is in affect. So we need to handle this case better in
indicating to the callers of cached_get() that the fetch has in fact
failed to due to the ratelimit.