backend/api/pull_request: Only calc changes with a single PR commit
The DB query returns all results matching a given pull request number
but we want to restrict things more than that. We need to only return
results for a specific git commit. Failure to do this leads to weird
things, e.g. where multiple change points are detected for a PR.