backend/db: Fix loop hole where invalid test results can persist
Since we introduced the change point cache, there are now two parts of
working with test results during HTTP POST:

 1. change points are calculated and cached
 2. test result data is added to the DB

We aren't performing any verification of the metric data during step 1,
which means test results with an invalid format can be added to the
change point cache but users will see the HTTP POST fail with a 400
status code at step 2, e.g. if the metric data is missing a unit field
which isn't used by the change point cache but is used when storing test
results.

Increase our verification of test result data

This change also necessitates updating a bunch of tests to adhere to the
now stricter rules for metrics.

Fixes #215