Also precompute summary data for the /summary end point
However, I only succeeded in doing this for the lowest level.
If you have multiple levels in the URI before  you get to the graph,
you will only see summaries at the last level before the actual graph.

The background.py tasks aren't scheduled by anything, they need to be triggered
by issuing a curl http request to /summary end point. This should be
fixed with something asyncio and like
https://stackoverflow.com/questions/73081130/python-fastapi-shedule-task

This approach is still slow, it still loops over every user and every test
and issues a read, which will trigger the computation (and caching) of change
points if needed. Ultimately we need to move to a more database centric
design, where for example in this case if we simply join/$lookup the cached
change point record, we can then use $exist to return the list of not cached
test results, and then compute those.