Fix: pymongo error "An empty update path is not valid."
This happened for for summaries because we use the test_name
path as the key in the summaries_cache dict/json. The root of
the hierarchy, ie the summary of all your tests, would have a
path of the empty string. This is not allowed by pymongo.

Fix is to skip computing the root / empty string as key. We don't
need it.