Add /admin/impersonate API and related frontend
A superuser can assume the identity of any other user. You continue
to use your JWT token. The Fastapi authentication layer overrides
this so you become the other user in the backend. In the frontend
localStorage('username') is set to the impersonated value and
localStorage('username_real') is the original, actual identity.
Note that the JWT token also includes your actual identity, and
this is the only thing sent to the backend.

In the frontend orangered CSS is used to bring attention to the
fact you are now using another users credentials.