Use incremental Hunter when possible
This is a bit tricky on Nyrkio level, because we write a new
test result immediately to MongoDB, then query back the full
PerformanceTestResultSeries. But at this point the new
data point looks the same as the old ones. But we proceed
by comparing the last_modified from the cached change points
and consider any test results newer than that as the new points
to append.