Merge different code paths for My, Org and Public Dashboards
This patch simply causes all three to call the same functions.
There is still way to many redundant functions that can be
removed now that all alternatives share the same code path.
But that will be for another day.

Reacht's useState() variables are apparently at their limit with
this amount of complexity. Made summaries work with a hack
for My Dashboards but Org and Public still don't have
the summaries printed.