First part of NoTokenHandshake
This is to enable all contributors to TursoDB to use Nyrkiö in
pull requests. Currently it only works for core developers that
have write access to the repo. By getting rid of nyrkio-token
completely (well, making it optional) we both simplify the user
experience and make all contributors equal.

In the first phase focus is on supporting public projects. Private
projects are possible to support, but paradoxically the repo
owner must first install Nyrkiö as a github app, so the benefit
is somewhat lost. (Still, no need to copy paste github-token.)

This commit only has the first part of the handshake.