Get correct contents for pr picker dropdown
The list of PRs to show for a given test results dashboard was
never correct in the first place. In addition allowing any github
user post results without jwt token now means that the set of
PRs to return no longer belong to a single user_id.

This adds a new MongoDB aggregation query that attempts to fetch
the right list from db

Note that MongoMotorMock continues to be broken for aggregate()
and will be tossed one of these days