Unify ChangePointSummaryTable
Previously, a different version of the table was used for leaves
vs intermediate nodes. The leaf node is still different in that
the column for test name is removed but other than that the same
code is shared everywhere. I think the commit messages are lost
with this commit, hopefully won't take many hours to copy the code
over to this table.

The fundamental part of this change is to actually embed and store
the direction information (if available) into the change point:
AnalyzedSeries in Hunter (Otava upgrade pending) and serialized
into the database. This is already provided for in Hunter via the
Metric() class, but we had not used it correctly and consistently
in Nyrkiö, and also the to_json() method provided by me did not
handle this correctly. So there's a corresponding Hunter
commit that this commit depends on.