Allow CPH users to trigger the notify=1 for PRs
...the common use case of course is to get the report for their own
PRs but actually anything in the same org is allowed.

Added elaborate tests as this was getting messy.

Fixed database queries so that the pull request is evaluated against
its target branch. Previously the PR would be stored with the
authors user id, and would therefore typically be its own data
point and wasn't really compared against the main branch'es
history of results. Now repo_owner is used to get the history.