Fix: The UserConfig class would reset core onfig to defaults
...as a side effect of setting any other config value.

The logic of the UserConfig class is that it is allowed to just
set some subset of the configuration. In such a scenario
the other fields are set to None. (Because they can't just not
exist like in a dict...) Then the DB class will ignore fields
that are None. However, UserConfig.core would always be set to
the default values 0.05 and 0.01 and since they aren't null
they would be stored in the database, overwriting any previous value.

Note that this does NOT happen if you actually set pvalue and threshold.
It DOES happen for example when setting notifier configuration.
(...something the nyrkio/change-detect github action dous.)