Separate the launching of ec2 instances into a background task
ec2 sdk operations, including things like ssh into the ec2 instance
are synchronous. So with two incoming webhooks requests the first would
block and timeout the second.

Finally also moves the /precompute to the worker01 container.

TODO: Start the container so that it's a true background worker,
looping forever and no http api.